requirements.txt
albucore==0.0.23
albumentations==2.0.5
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.34.1
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.15.2
setuptools==78.1.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
streamlit==1.44.1
stringzilla==3.12.3
sympy==1.13.1
tenacity==9.1.2
timm==1.0.15
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
